Add apiCall to App and create a dispatch in form for the word
